Change emulator flags to uint32

+++ inc/emulator.h
@@ -42,7 +42,7 @@ extern "C" {
 /*------------------------------------------------------------------------------
  Macros  
 ------------------------------------------------------------------------------*/
-typedef uint8_t EMULATOR_FLAGS_TYPE; 
+typedef uint32_t EMULATOR_FLAGS_TYPE; 
 
 #define IGNITE_FLAG_BIT (1 << 0)
 #define IRQ_ENABLED_FLAG_BIT (1 << 1)
